Remove org files from the distro
Since Foliot is part of GNU and has a bug-foliot mailing list, there is
no real need for a foliot-bugs.org file anymore.

The foliot.org is/was supposed to be the project TODO file, but having
it under git and distributed sort of refrained me using it 'properly',
so I decided to maintain it locally.

* org/foliot-bugs.org:
* org/foliot.org:  Deleted.

* Makefile.am: Updated for the above.